# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Product:
    properties:
      description:
        maxLength: 500
        minLength: 10
        type: string
      id:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - description
    - name
    - price
    - stock
    type: object
  domain.User:
    properties:
      access_token:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  request.CreateProductRequestDTO:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - description
    - name
    - price
    - stock
    type: object
  request.CreateUserRequestDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  request.LoginRequestDTO:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  response.AuthResponseDTO:
    properties:
      access_token:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Guilherme Azevedo
  description: API REST da Demostore, com gerenciamento de usuários, autenticação
    e produtos.
  title: Demostore API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Realiza login com email e senha, retornando um access token JWT
      parameters:
      - description: Credenciais de login
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AuthResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Autentica um usuário
      tags:
      - auth
  /products:
    get:
      description: Retorna todos os produtos cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Product'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lista todos os produtos
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Cadastra um novo produto na plataforma
      parameters:
      - description: Dados do produto
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/request.CreateProductRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cria um novo produto
      tags:
      - products
  /products/{id}:
    get:
      description: Retorna os dados de um produto a partir do seu UUID
      parameters:
      - description: ID do produto (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Busca um produto pelo ID
      tags:
      - products
  /users:
    post:
      consumes:
      - application/json
      description: Cadastra um novo usuário na plataforma
      parameters:
      - description: Dados do usuário
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Cria um novo usuário
      tags:
      - users
  /users/{id}:
    get:
      description: Retorna os dados de um usuário a partir do seu UUID
      parameters:
      - description: ID do usuário (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Busca um usuário pelo ID
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Informe o token no formato: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"